Applied font change, added Tip text
I felt we deeply needed a font upgrade. I imported "ink free" which is in LibreOffice and comes with Windows. I applied it to 'Anxiety' 'Panic attack' and 'NO!".
How about we add the control for Stamina to the gameover screen? They struggle at least once and feel the game before they get to soar through? See screenshot on Discord